backend/requirements.txt
# CareerForge Backend Dependencies
# Google AI / ADK
google-adk>=1.0.0
google-genai>=1.0.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=13.0
sse-starlette>=2.0.0

# Data Models & Config
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# PDF Generation
reportlab>=4.0.0

# HTTP Client (used by genai SDK)
httpx>=0.27.0
